# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SatQuery API
  description: |
    Earth Observation intelligence API. Upload satellite imagery, draw regions,
    ask questions about specific areas, and get answers backed by visual evidence
    and impact statistics.
  version: "1.0.0"
  contact:
    name: SatQuery Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: "{API_BASE_URL}"
    description: Configurable base URL
    variables:
      API_BASE_URL:
        default: http://localhost:8000
...
paths:
...
  /images/upload:
    post:
      operationId: uploadImage
      tags: [images]
      summary: Upload a satellite image
      description: |
        Accepts a GeoTIFF or standard image file. Extracts metadata (CRS, bounds,
        resolution, sensor, band count) using rasterio. Returns image_id and
        extracted metadata. For non-GeoTIFF images, spatial metadata fields will
        be null.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: GeoTIFF or image file (PNG/JPEG/TIFF)
                sensor:
                  type: string
                  description: Optional sensor override (e.g. "Sentinel-2", "SAR")
                description:
                  type: string
                  description: Optional description for the image
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageMetadata"
        "400":
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{image_id}:
    get:
      operationId: getImage
      tags: [images]
      summary: Get image metadata
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: Image metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageMetadata"
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{image_id}/tile:
    get:
      operationId: getImageTile
      tags: [images]
      summary: Get image as a web-viewable tile (PNG)
      description: |
        Returns the uploaded image converted to PNG for frontend display.
        Optionally crop to a bounding box in pixel coordinates.
      parameters:
        - $ref: "#/components/parameters/ImageId"
        - name: x_min
          in: query
          schema: { type: integer }
        - name: y_min
          in: query
          schema: { type: integer }
        - name: x_max
          in: query
          schema: { type: integer }
        - name: y_max
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Image tile
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: Image not found

  # ─── Regions ─────────────────────────────────────────────
...
  /query:
    post:
      operationId: submitQuery
      tags: [query]
      summary: Ask a question about an image region
      description: |
        Sends a natural-language question about an image (optionally scoped to
        a region). The backend routes to the appropriate model based on context:
        - Single image → GeoChat VQA
        - Region specified → crops image first, then VQA
        
        Supports SSE streaming via Accept: text/event-stream header.
        When streaming, events are sent as:
        - event: trace, data: {step_name, status, duration_ms}
        - event: token, data: {text}
        - event: evidence, data: {type, bbox, confidence, label}
        - event: done, data: {query_id, confidence}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query response (non-streaming)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
            text/event-stream:
              schema:
                type: string
                description: SSE stream of trace/token/evidence/done events
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image or region not found
  /query/{query_id}:
    get:
      operationId: getQuery
      tags: [query]
      summary: Get a past query result
      parameters:
        - $ref: "#/components/parameters/QueryId"
      responses:
        "200":
          description: Query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "404":
          description: Query not found

  # ─── Temporal (Change Detection) ────────────────────────
...
components:
  parameters:
    ImageId:
      name: image_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QueryId:
      name: query_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, detail]
      properties:
        error:
          type: string
        detail:
          type: string

    # ─── Image ───────────────────────────────────────────
    ImageMetadata:
      type: object
      required: [id, filename, file_path, upload_time]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_path:
          type: string
          description: Internal storage path
        crs:
          type: string
          nullable: true
          description: "Coordinate reference system (e.g. EPSG:4326)"
        bounds:
          $ref: "#/components/schemas/GeoJsonPolygon"
          nullable: true
        resolution_m:
          type: number
          format: double
          nullable: true
          description: Ground sample distance in meters
        sensor:
          type: string
          nullable: true
        band_count:
          type: integer
          nullable: true
        width_px:
          type: integer
          nullable: true
        height_px:
          type: integer
          nullable: true
        metadata_json:
          type: object
          nullable: true
          description: Additional rasterio metadata
        upload_time:
          type: string
          format: date-time

    # ─── Region ──────────────────────────────────────────
...
    PixelBounds:
      type: object
      required: [x_min, y_min, x_max, y_max]
      properties:
        x_min:
          type: integer
        y_min:
          type: integer
        x_max:
          type: integer
        y_max:
          type: integer

    # ─── Query ───────────────────────────────────────────
    QueryRequest:
      type: object
      required: [image_id, question]
      properties:
        image_id:
          type: string
          format: uuid
        region_id:
          type: string
          format: uuid
          nullable: true
          description: Optional region to scope the question to
        question:
          type: string
          description: Natural-language question about the image/region
    QueryResponse:
      type: object
      required: [query_id, answer, confidence, evidence, execution_trace]
      properties:
        query_id:
          type: string
          format: uuid
        answer:
          type: string
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        execution_trace:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionTraceStep"
        model_used:
          type: string
          nullable: true
    Evidence:
      type: object
      required: [type, confidence]
      properties:
        type:
          type: string
          enum: [bbox, mask, point]
        bbox:
          $ref: "#/components/schemas/PixelBounds"
          nullable: true
        mask_url:
          type: string
          nullable: true
          description: URL to download the mask image
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        label:
          type: string
          nullable: true
          description: "Object class label (e.g. building, road, flood)"
    ExecutionTraceStep:
      type: object
      required: [step_name, status, duration_ms]
      properties:
        step_name:
          type: string
          description: "e.g. Validating region, Cropping image, Running VQA"
        status:
          type: string
          enum: [pending, running, completed, failed]
        duration_ms:
          type: integer
        detail:
          type: string
          nullable: true

    # ─── Temporal ────────────────────────────────────────
...
    GeoJsonPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          description: Array of linear rings
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 3